// apps/dapp/src/global/enum.tsx
export enum ButtonType {
  Primary = 'primary',
  Secondary = 'secondary',
  Accent = 'accent',
  Ghost = 'ghost',
  Success = 'success',
  Error = 'error',
  Link = 'link',
}

export enum JobType {
  CoFounder = 'cofounder',
  CEO = 'ceo',
  Dev = 'dev',
  CMO = 'cmo',
  COO = 'coo',
  CM = 'cm',
  Advisor = 'advisor',
  GMX = 'gmx',
}

export enum NavigationCaterogy {
  Home,
  Project,
  Ressources,
}

export enum GBC_ADDRESS {
  ZERO = '0x0000000000000000000000000000000000000000',
  GBC = '0x17f4baa9d35ee54ffbcb2608e20786473c7aa49f',
  POLICE = '0x575F40E8422EfA696108dAFD12cD8d6366982416',
  LAB = '0xF4f935F4272e6FD9C779cF0036589A63b48d77A7',
  PROFILE = '0xeab999593077ab4c799299C5CAa4f140Cf5fa27E',
  CLOSET = '0xb3598b108F59AfEc664974DfD96330DE48eF2D4D',
  GOVERNOR = '0x1F1da4CDDAf2A12f9f182D9CD6c4607E2A649105',

  WETH = '0xEBbc3452Cc911591e4F18f3b36727Df45d6bd1f9',

  TREASURY_ARBITRUM = '0xde2dbb7f1c893cc5e2f51cbfd2a73c8a016183a0',
  TREASURY_AVALANCHE = '0x753b4769154fd100ee763e927305d5b3131dbc8e',
}

export enum BannerStatus {
  Success,
  Error,
  Info,
}

export enum CategoryIndex {
  Background = 0,
  Shirt = 1,
  Smile = 3,
  Mask = 4,
  Hat = 5,
  Star = 6,
}

export enum LabItemGroupType {
  Wearable = 0,
  Face = 1,
  Special = 2,
  Background = 3,
}